+++ Swift Vocabulary/Info.plist
