+++ Swift Vocabulary/Resources/AppDelegate.swift
@@ -2,14 +2,13 @@
 //  AppDelegate.swift
 //  Swift Vocabulary
 //
-//  Created by Rohan Taylor on 11/7/19.
+//  Created by Rohan Taylor on 12/9/19.
 //  Copyright © 2019 Rohan Taylor. All rights reserved.
 //
 
 import UIKit
 
 @UIApplicationMain
-
 class AppDelegate: UIResponder, UIApplicationDelegate {
