+++ Swift Vocabulary/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
