+++ Swift Vocabulary/Resources/Assets.xcassets/Contents.json
