+++ Swift Vocabulary/Resources/Base.lproj/LaunchScreen.storyboard
