+++ Swift Vocabulary/Resources/SceneDelegate.swift
@@ -2,7 +2,7 @@
 //  SceneDelegate.swift
 //  Swift Vocabulary
 //
-//  Created by Rohan Taylor on 11/7/19.
+//  Created by Rohan Taylor on 12/9/19.
 //  Copyright © 2019 Rohan Taylor. All rights reserved.
 //
