+++ Swift Vocabulary/VocabularyWord.swift
@@ -2,14 +2,13 @@
 //  VocabularyWord.swift
 //  Swift Vocabulary
 //
-//  Created by Rohan Taylor on 11/7/19.
+//  Created by Rohan Taylor on 12/9/19.
 //  Copyright © 2019 Rohan Taylor. All rights reserved.
 //
 
-import Foundation
-
+import UIKit
 
 struct VocabularyWord {
-    var word : String
-    var definition : String
+   let word : String
+   let definition : String
 }
