+++ Swift Vocabulary/WorldsTableTableViewController.swift
@@ -1,0 +1,102 @@
+//
+//  WorldsTableTableViewController.swift
+//  Swift Vocabulary
+//
+//  Created by Rohan Taylor on 12/9/19.
+//  Copyright © 2019 Rohan Taylor. All rights reserved.
+//
+
+import UIKit
+
+class WorldsTableTableViewController: UITableViewController {
+
+    
+    var vocabWords: [VocabularyWord] = [VocabularyWord(word: "Rohan", definition: "Ascending"),VocabularyWord(word: "Is", definition: "third person singular present of be."),VocabularyWord(word: "Awesome", definition: "extremely impressive or daunting; inspiring great admiration, apprehension, or fear.")]
+
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Uncomment the following line to preserve selection between presentations
+        // self.clearsSelectionOnViewWillAppear = false
+
+        // Uncomment the following line to display an Edit button in the navigation bar for this view controller.
+        // self.navigationItem.rightBarButtonItem = self.editButtonItem
+    }
+
+    // MARK: - Table view data source
+
+
+    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        // #warning Incomplete implementation, return the number of rows
+        return vocabWords.count
+    }
+
+    override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        let cell = tableView.dequeueReusableCell(withIdentifier: "WordCell", for: indexPath)
+        
+        let vocabWord = vocabWords[indexPath.row]
+        
+        cell.textLabel?.text = vocabWord.word
+
+        // Configure the cell...
+
+        return cell
+    }
+    
+
+    /*
+    // Override to support conditional editing of the table view.
+    override func tableView(_ tableView: UITableView, canEditRowAt indexPath: IndexPath) -> Bool {
+        // Return false if you do not want the specified item to be editable.
+        return true
+    }
+    */
+
+    /*
+    // Override to support editing the table view.
+    override func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCell.EditingStyle, forRowAt indexPath: IndexPath) {
+        if editingStyle == .delete {
+            // Delete the row from the data source
+            tableView.deleteRows(at: [indexPath], with: .fade)
+        } else if editingStyle == .insert {
+            // Create a new instance of the appropriate class, insert it into the array, and add a new row to the table view
+        }    
+    }
+    */
+
+    /*
+    // Override to support rearranging the table view.
+    override func tableView(_ tableView: UITableView, moveRowAt fromIndexPath: IndexPath, to: IndexPath) {
+
+    }
+    */
+
+    /*
+    // Override to support conditional rearranging of the table view.
+    override func tableView(_ tableView: UITableView, canMoveRowAt indexPath: IndexPath) -> Bool {
+        // Return false if you do not want the item to be re-orderable.
+        return true
+    }
+    */
+
+
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        
+        if segue.identifier == "ShowDefinitionSegue" {
+            
+            if let indexPath = tableView.indexPathForSelectedRow,
+                let definitionVC = segue.destination as? DefinitionViewController {
+                
+                let vocabWord = vocabWords[indexPath.row]
+                definitionVC.vocabWord = vocabWord
+            }
+        }
+        // Get the new view controller using segue.destination.
+        // Pass the selected object to the new view controller.
+    }
+
+}
